requirements.txt
albumentations==1.1.0
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
bokeh==2.3.0
Bottleneck==1.3.2
branca==0.4.2
Brotli==1.0.9
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
dash==1.19.0
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
decorator==4.4.2
defusedxml==0.6.0
ecdsa==0.15
efficientnet-pytorch==0.6.3
entrypoints==0.3
esptool==2.8
Flask==1.1.2
Flask-Compress==1.9.0
folium==0.12.1
future==0.18.2
graphviz==0.16
greenlet==1.1.2
h5py==2.10.0
htmlmin==0.1.12
hydra-core==1.1.1
idna==2.10
ImageHash==4.2.1
imageio==2.10.3
ipykernel==5.4.3
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
matplotlib-inline==0.1.3
missingno==0.5.0
mistune==0.8.4
monai==0.8.0
mpmath==1.2.1
multimethod==1.4
munch==2.5.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.6.3
nose==1.3.7
notebook==6.2.0
numpy==1.20.0
omegaconf==2.1.1
opencv-contrib-python==4.5.1.48
opencv-python==4.5.4.58
opencv-python-headless==4.5.4.58
packaging==20.9
pandas==1.2.1
pandas-profiling @ https://github.com/pandas-profiling/pandas-profiling/archive/refs/tags/v3.0.0.zip
pandasql==0.7.3
pandocfilters==1.4.3
parso==0.8.1
patsy==0.5.2
phik==0.12.0
pickleshare==0.7.5
Pillow==8.4.0
plotly==4.14.3
pretrainedmodels==0.7.4
prometheus-client==0.9.0
prompt-toolkit==3.0.14
py4j==0.10.9.2
pyaes==1.6.1
pycparser==2.20
pydantic==1.8.2
Pygments==2.7.4
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyrsistent==0.17.3
pyserial==3.4
pyspark==3.2.0
python-dateutil==2.8.1
pytorch-toolbelt==0.4.4
pytz==2021.1
PyWavelets==1.1.1
#pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.0.2
qtconsole==5.1.0
QtPy==1.9.0
qudida==0.0.4
requests==2.25.1
retrying==1.3.3
scikit-image==0.18.3
scikit-learn==0.24.1
scipy==1.6.0
seaborn==0.11.1
segmentation-models-pytorch==0.2.0
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
SQLAlchemy==1.4.26
statsmodels==0.13.1
sympy==1.8
tangled-up-in-unicode==0.1.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2021.11.2
timm==0.4.12
torch==1.10.0
torchaudio==0.10.0
torchgeometry==0.1.2
torchmetrics==0.6.0
torchvision==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.3
visions==0.7.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
